Add small static file

+++ i
@@ -1770,10 +1770,15 @@ confirm_import_db() {
 }
 
 import_static() {
-    wget -O /home/images.tar http://static.cinemapress.org/images.tar
-    mkdir -p /var/local/images/poster
-    wget http://cinemapress.org/images/web/no-poster.gif -qO /var/local/images/poster/no-poster.gif
-    wget http://cinemapress.org/images/web/no-poster.jpg -qO /var/local/images/poster/no-poster.jpg
+    if [ ! -f "/var/local/images/poster/no-poster.jpg" ]
+    then
+        wget -O /home/images.tar http://static.cinemapress.org/images.tar
+        mkdir -p /var/local/images/poster
+        wget http://cinemapress.org/images/web/no-poster.gif -qO /var/local/images/poster/no-poster.gif
+        wget http://cinemapress.org/images/web/no-poster.jpg -qO /var/local/images/poster/no-poster.jpg
+    else
+        wget -O /home/images.tar http://static.cinemapress.org/last.tar
+    fi
     printf "\n${NC}"
     printf "${G}Распаковка в фоновом режиме ...\n"
     printf "${NC}Может занять несколько часов.\n"
